fix API crud project and issue, add API get issue and project by id

+++ app/api/project/[id]/route.js
@@ -2,22 +2,61 @@ import { updateDoc, serverTimestamp, getDoc, deleteDoc, doc } from 'firebase/fir
 import { NextResponse } from "next/server";
 import { db } from '@/app/firebase/config';
 
-export async function PUT(request, { params }) {
+export async function GET(request, context) {
     try {
-        const { id }  = params;
-        const { key, projectName } = await request.json();
+        const { id }  = context.params;
 
         const projectDocRef = doc(db, 'projects', id);
 
-        console.log(projectDocRef)
+        if(!projectDocRef) {
+            return  NextResponse.json({
+                message: "Project not found"
+            }, { status: 404 });
+        }
+
+        const projectSnap = await getDoc(projectDocRef);
+        const projectData = projectSnap.data()
+
+        if(!projectData) {
+            return  NextResponse.json({
+                message: "Project detail not found"
+            }, { status: 404 });
+        }
+
+        return  NextResponse.json({
+            data: {
+                id: projectSnap.id,
+                ...projectData
+            },
+            message: "Successfully retrieved project"
+        }, { status: 404 });
+
+    } catch (error) {
+        console.error("Cannot get project detail", error);
+        return NextResponse.json({
+            data: null,
+            message: error.message
+        }, { status: 500 });
+    }
+}
+
+export async function PUT(request, context) {
+    try {
+        const { id }  = context.params;
+        const { key, projectName } = await request.json();
+
+        const projectDocRef = doc(db, 'projects', id);
+        const projectSnap = await getDoc(projectDocRef);
+        const projectData = projectSnap.data()
 
         await updateDoc(projectDocRef, {
-            key: key,
-            projectName: projectName,
+            key: key ?? projectData.key,
+            projectName: projectName ?? projectData.projectName,
             updatedAt: serverTimestamp()
         });
 
         const updatedProjectSnap = await getDoc(projectDocRef);
+
         if (updatedProjectSnap.exists()) {
             return NextResponse.json({
                 data: {
@@ -43,10 +82,16 @@ export async function PUT(request, { params }) {
     }
 }
 
-export async function DELETE({ params }) {
+export async function DELETE(request, context) {
     try {
-        const { id } = params;
+        const { id } = context.params;
         const projectDocRef = doc(db, 'projects', id);
+
+        if (!projectDocRef) {
+            return NextResponse.json({
+                message: "Project not found"
+            }, { status: 404 })
+        }
 
         await deleteDoc(projectDocRef);
